+++ src/js/components/refs.js
@@ -1,0 +1,8 @@
+const refs = {
+    galleryItem: document.querySelectorAll('.galleryItem'),
+    modalContent: document.querySelector('#modal-1'),
+    addToWatchedBtn: document.querySelector('[data-attribute="add-to-watched"]'),
+    addToQueueBtn: document.querySelector('[data-attribute="add-to-queue"]')
+};
+
+export default refs;
