--- src/js/refs.js
@@ -1,10 +1,0 @@
-const refs = {
-    galleryItem: document.querySelector("[data-index='galleryItem']"),
-    modalContent: document.querySelector('#modal-1-content'),
-    modalOverlay: document.querySelector('.modal__overlay'),
-    modalCloseButton: document.querySelector('.modal__close'),
-    addToWatchedBtn: document.querySelector('[data-attribute="add-to-watched"]'),
-    addToQueueBtn: document.querySelector('[data-attribute="add-to-queue"]')
-};
-
-export default refs;
