--- src/modal.html
@@ -1,32 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>modal window</title>
-</head>
-<body>
-<!-- modal window -->
-<section>
-    <div id="modal-1" aria-hidden="true">
-
-        <!-- [2] -->
-        <div tabindex="-1" data-micromodal-close>
-
-            <!-- [3] -->
-            <div role="dialog" aria-modal="true" aria-labelledby="modal-1-title">
-
-
-                <header>
-                    <!-- [4] -->
-                    <button aria-label="Close modal" data-micromodal-close></button>
-                </header>
-
-                <div id="modal-1-content"></div>
-
-            </div>
-        </div>
-    </div>
-</section>
-</body>
-</html>
