--- src/templates/insideModalMarkup.hbs
@@ -1,25 +1,0 @@
-<img src="https://image.tmdb.org/t/p/w300{{poster_path}}" alt="poster of {{original_title}}">
-<div>
-    <h2 class="movie-title">{{original_title}}</h2>
-    <p class="movie-item">
-        <i class="movie-item-title">Vote/Votes</i>
-        <span class="value">{{vote_average}} / {{vote_count}}</span>
-    </p>
-    <p class="movie-item">
-        <i class="movie-item-title">Popularity</i>
-        <span class="value">{{popularity}}</span>
-    </p>
-    <p class="movie-item">
-        <i class="movie-item-title">Original Title</i>
-        <span class="value">{{original_title}}</span>
-    </p>
-    <p class="movie-item">
-        <i class="movie-item-title">Genre</i>
-        <span class="value">{{genre_ids}}</span>
-    </p>
-    <article class="movie-review">About <br> {{overview}}</article>
-</div>
-<div>
-    <button class="btn" data-attribute="add-to-watched">ADD TO WATCHED</button>
-    <button class="btn" data-attribute="add-to-queue">ADD TO QUEUE</button>
-</div>
